+++ NEWS.md
@@ -1,5 +1,29 @@
 
 <!-- NEWS.md is generated from NEWS.Rmd. Please edit that file -->
+eq5d 0.2.0 (2019-06-27)
+=======================
+
+-   Added Brazilian, Canadian, Chilean, Polish, Portuguese, Taiwanese and Trinidad and Tobago EQ-5D-3L TTO value sets.
+
+-   Added Taiwanese EQ-5D-5L VT value set.
+
+-   Example data added to shiny app.
+
+-   Plots tab and ggplot2 density and ecdf plots added to shiny app.
+
+-   Added functionality to shiny app to calculate a single index score.
+
+-   Notification added to shiny app to warn when EQ-5D-5L is selected, but all uploaded values are 1, 2 or 3.
+
+eq5d 0.1.1 (2019-05-27)
+=======================
+
+-   Fixed issue where displayed description values were mislabelled in the Shiny app table when the columns in the uploaded file weren't in the order MO, SC, UA, PD, AD.
+
+-   Fixed issue with Shiny app download button not working correctly.
+
+-   Added checkbox to Shiny app to give the option of displaying the index scores with all info provided in the uploaded file or just with the dimension scores.
+
 eq5d 0.1.0 (2019-05-10)
 =======================